arkui-x update master OH to weekly_20260105 and OH fix build bug

Created-by: mufeng_413
Commit-by: zhangqing
Merged-by: ohci1
Description: https://gitcode.com/openharmony/multimedia_image_framework/issues/3994

See merge request: openharmony/third_party_ffmpeg!578

+++ libavcodec/ffjni.h
@@ -24,7 +24,6 @@
 #define AVCODEC_FFJNI_H
 
 #include <jni.h>
-#include <stddef.h>
 
 /*
  * Attach permanently a JNI environment to the current thread and retrieve it.
@@ -106,7 +105,7 @@ struct FFJniField {
     const char *method;
     const char *signature;
     enum FFJniFieldType type;
-    size_t offset;
+    int offset;
     int mandatory;
 
 };
